# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CodexDominion Civilization API
  version: 1.0.0
  description: >
    Core API for CodexDominion 2.0 — youth, creators, circles, missions,
    culture, ambassadors, events, and analytics.
servers:
  - url: https://api.codexdominion.com/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Mission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        season_id:
          type: string
          format: uuid
        month:
          type: integer
        week:
          type: integer
        type:
          type: string
          enum: [GLOBAL, REGIONAL, CIRCLE]
...
    Circle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        captain_id:
          type: string
          format: uuid
        region_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
...
    CulturalStory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: object
          properties:
            format:
              type: string
              example: markdown
            body:
              type: string
        season_id:
          type: string
          format: uuid
          nullable: true
        week:
          type: integer
          nullable: true
        region_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
    Ritual:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [OPENING, CLOSING, SEASONAL, UNITY, CEREMONY]
    Region:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        country:
          type: string
        timezone:
          type: string
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        event_type:
          type: string
          enum: [LAUNCH, CEREMONY, SUMMIT, TRAINING, OTHER]
        region_id:
          type: string
          format: uuid
          nullable: true
        scheduled_at:
          type: string
          format: date-time
    AnalyticsOverview:
      type: object
      properties:
        active_youth:
          type: integer
        active_circles:
          type: integer
        missions_completed_this_week:
          type: integer
        regions_active:
          type: integer
security:
  - bearerAuth: []
paths:
...
  /missions:
    get:
      summary: List missions
      tags: [Missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: season_id
          schema:
            type: string
            format: uuid
        - in: query
          name: month
          schema:
            type: integer
      responses:
        '200':
          description: Missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mission'
    post:
      summary: Create mission
      tags: [Missions]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, season_id, type]
              properties:
                title:
                  type: string
                description:
                  type: string
                season_id:
                  type: string
                  format: uuid
                month:
                  type: integer
                week:
                  type: integer
                type:
                  type: string
                  enum: [GLOBAL, REGIONAL, CIRCLE]
      responses:
        '201':
          description: Mission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
...
  /circles:
    get:
      summary: List circles
      tags: [Circles]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Circles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Circle'
    post:
      summary: Create circle
      tags: [Circles]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, captain_id]
              properties:
                name:
                  type: string
                captain_id:
                  type: string
                  format: uuid
                region_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Circle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
...
  /culture/story/current:
    get:
      summary: Get current cultural story
      tags: [Culture]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current cultural story
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CulturalStory'
  /culture/stories:
    get:
      summary: List cultural stories
      tags: [Culture]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: season_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cultural stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CulturalStory'
...
  /culture/rituals:
    get:
      summary: List rituals
      tags: [Culture]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Rituals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ritual'

  # ----------------------------------------------------
  # 8. AMBASSADOR & EXPANSION
  # ----------------------------------------------------
  /regions:
    get:
      summary: List regions
      tags: [Regions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
  /regions/{id}:
    get:
      summary: Get region
      tags: [Regions]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Region
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
...
  /events:
    get:
      summary: List events
      tags: [Events]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: region_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create event
      tags: [Events]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, event_type, scheduled_at]
              properties:
                title:
                  type: string
                description:
                  type: string
                event_type:
                  type: string
                  enum: [LAUNCH, CEREMONY, SUMMIT, TRAINING, OTHER]
                region_id:
                  type: string
                  format: uuid
                scheduled_at:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
...
  /analytics/overview:
    get:
      summary: Get high-level analytics overview
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'